img_processing/deployment/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorzero==2.0
flatbuffers==23.5.26
future==0.18.3
gast==0.5.4
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
gpiozero==2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.0
iso8601==2.1.0
keras==2.15.0
lgpio==0.0.0.2
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==9.5.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyserial==3.5
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
RPi-GPIO-i2c-LCD==0.1.3
RPi.GPIO==0.7.1
rsa==4.9
serial==0.0.97
six==1.16.0
smbus==1.1.post2
smbus2==0.4.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
typing-extensions==4.8.0
urllib3==2.1.0
werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
